Patch: make latencies depend on inputs (2 comps)
Apply a patch that changes how the latencies are calculated for some
components, making them depend on the values of the inputs.
These components are And and Multiplexer, for now.

The objective is, in the end, to make the critical path displayed to
depend on the instruction currently being executed (critical path of
the instruction).