Calculate CPU performance (for statistics)
Calculate both the instruction performance and CPU performance, for the
statistics dialog.

While testing, I noticed that in the unicycle.cpu datapath, while
executing a lw instruction, the CPU clock period is 930 ps but the
maximum accumulated latency of all inputs is 915 ps (in
REG.WriteData input).
This is due to MUX_REG's accumulated latency being 365 ps instead of 415
ps during a lw.