Add property to change the displayed "Usage" help
The help text will contain a "Usage" line. The default value of the line
is:
  Usage: java -jar DrMIPS.jar [options] [file]
If the simulator will be started from a launcher script/program, you may
want to change the displayed "Usage" line. To do that, you must define
the "program.name" property in the launcher script, like so:
  java -Dprogram.name=drmips -jar /path/to/DrMIPS.jar
This example will change the "Usage" line to:
  Usage: drmips [options] [file]