Improve PT translation of "assemble" and related
"Assemblado" sounds strange and I think it's not a Portuguese word.
"Assembled code" was translated to "Código máquina" and "assemble" was
translated to "gerar código máquina" (adapted to each case). They still
sound a bit strange, but are more correct.
The user manual was also updated.