Fix bug in pipeline.cpu and pipeline-extended.cpu.
These pipelines predict branches as not taken. In the MEM stage if it
turns out that the branch should have been taken the earlier pipeline
stages are flushed and the PC is updated to the correct value.

However if the Hazzard detection unit detects a hazzard at this time, it
disables writes to the PC, and therefore the PC never gets updated, and
so the branch is fully not taken.

To fix this you must ignore hazzards if the MEM stage has determined
that the branch was actually taken.